Refactor email (SendGrid) contact logs
